Made delete functionality admin-only.

+++ delete/cancer.php
@@ -34,40 +34,39 @@ $iv_query= mysqli_fetch_assoc(mysqli_query($connect, "select riv from norm"));
 $iv=$iv_query['riv'];
 
 // ID encrypted
-//$enc_id=openssl_encrypt($id, $cipher, $encryption_key, 0, $iv);
 $enc_id="0x".bin2hex(openssl_encrypt($id, $cipher, $encryption_key, 0, $iv));
 
 mysqli_close($connect);
 
-try {
-	$sql = "DELETE FROM `Diseases`
-	WHERE
-		`id` = $enc_id
-		AND `date` = '$date'
-		AND `type` = '$type'
-		AND `histology` = '$histology'
-		AND `status` = '$status'
-		AND `code` = '$location'
-		AND `side` = '$side'
-		AND `oncotree` = '$oncotree'
-		AND `clinicalsg` = '$clinicalsg'
-		AND `clinicalss` = '$clinicalss'
-		AND `pathologicsg` = '$pathologicsg'
-		AND `pathologicss` = '$pathologicss'
-		AND `comments` = '$comment'";
+$hasAdminRole = in_array("admin", explode(",", strtolower($roles)));
 
-	$sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
-	VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
+if ($hasAdminRole) {
+    $sql = "DELETE FROM `Diseases`
+    WHERE
+        `id` = $enc_id
+        AND `date` = '$date'
+        AND `type` = '$type'
+        AND `histology` = '$histology'
+        AND `status` = '$status'
+        AND `code` = '$location'
+        AND `side` = '$side'
+        AND `oncotree` = '$oncotree'
+        AND `clinicalsg` = '$clinicalsg'
+        AND `clinicalss` = '$clinicalss'
+        AND `pathologicsg` = '$pathologicsg'
+        AND `pathologicss` = '$pathologicss'
+        AND `comments` = '$comment'";
 
-	if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
-		echo "Success";
-	} else {
-		$error = mysqli_error($conn);
-		echo "There was a problem while deleting the data.";
-		echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
-	}
-} catch (Exception $e) {
-	echo "The data could not be deleted. Please contact the administrators.";
+    $sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
+    VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
+
+    if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
+        echo "Success";
+    } else {
+        $error = mysqli_error($conn);
+        echo "There was a problem while deleting the data.";
+        echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
+    }
 }
 
 mysqli_close($conn);
