+++ delete/cbc.php
@@ -1,48 +1,52 @@
 <?php
-	include('../configuration/db.php');
-	include('../configuration/mcode.php');
-	include('../configuration/key.php');
 
-	// Ip address of the user
-	$ip=$_POST['ip'];
-	$datesystem=$_POST['datesystem'];
-	$email=$_POST['email'];
-	$username=$_POST['username'];
-	$roles=$_POST['roles'];
-	$tracking=$_POST['tracking'];
+include('../configuration/db.php');
+include('../configuration/mcode.php');
+include('../configuration/key.php');
 
-	$id=$_POST['id'];
-	$date=$_POST['date'];
-	$type=$_POST['type'];
+// Ip address of the user
+$ip=$_POST['ip'];
+$datesystem=$_POST['datesystem'];
+$email=$_POST['email'];
+$username=$_POST['username'];
+$roles=$_POST['roles'];
+$tracking=$_POST['tracking'];
 
-	//Encryption
-	$encryption_key = hex2bin($key);
+$id=$_POST['id'];
+$date=$_POST['date'];
+$type=$_POST['type'];
 
-	// initialization vector
-	$iv_query= mysqli_fetch_assoc(mysqli_query($connect, "select riv from norm"));
-	$iv=$iv_query['riv'];
+//Encryption
+$encryption_key = hex2bin($key);
 
-	// ID encrypted
-	$enc_id="0x".bin2hex(openssl_encrypt($id, $cipher, $encryption_key, 0, $iv));
+// initialization vector
+$iv_query= mysqli_fetch_assoc(mysqli_query($connect, "select riv from norm"));
+$iv=$iv_query['riv'];
 
-	mysqli_close($connect);
+// ID encrypted
+$enc_id="0x".bin2hex(openssl_encrypt($id, $cipher, $encryption_key, 0, $iv));
 
-	$sql = "DELETE FROM `CBC`
-		WHERE
-			`id` = $enc_id
-			AND `date` = '$date'
-			AND	`type` = '$type'";
+mysqli_close($connect);
 
-	$sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
-	VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
+$hasAdminRole = in_array("admin", explode(",", strtolower($roles)));
 
-	if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
-		echo "Success";
-	} else {
-		$error = mysqli_error($conn);
-		echo "There was a problem while deleting the data. ";
-		echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
-	}
+if ($hasAdminRole) {
+    $sql = "DELETE FROM `CBC`
+        WHERE
+            `id` = $enc_id
+            AND `date` = '$date'
+            AND	`type` = '$type'";
 
+    $sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
+    VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
 
-	mysqli_close($conn);
+    if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
+        echo "Success";
+    } else {
+        $error = mysqli_error($conn);
+        echo "There was a problem while deleting the data. ";
+        echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
+    }
+}
+
+mysqli_close($conn);
