+++ delete/labs.php
@@ -32,29 +32,32 @@ $iv=$iv_query['riv'];
 //$enc_id=openssl_encrypt($id, $cipher, $encryption_key, 0, $iv);
 $enc_id="0x".bin2hex(openssl_encrypt($id, $cipher, $encryption_key, 0, $iv));
 
+$hasAdminRole = in_array("admin", explode(",", strtolower($roles)));
+
 mysqli_close($connect);
 
-$sql = "DELETE FROM `Lab`
-    WHERE
-        `id` = $enc_id
-        AND `date` = '$date'
-        AND `location` = '$location'
-        AND `height` = '$height'
-        AND `weight` = '$weight'
-        AND `diastolic` = '$diastolic'
-        AND `systolic` = '$systolic'
-        AND `comment` = '$comment'";
+if ($hasAdminRole) {
+    $sql = "DELETE FROM `Lab`
+        WHERE
+            `id` = $enc_id
+            AND `date` = '$date'
+            AND `location` = '$location'
+            AND `height` = '$height'
+            AND `weight` = '$weight'
+            AND `diastolic` = '$diastolic'
+            AND `systolic` = '$systolic'
+            AND `comment` = '$comment'";
 
-$sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
-VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
+    $sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
+    VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
 
-if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
-    echo "Success";
-} else {
-    $error = mysqli_error($conn);
-    echo "There was a problem while deleting the data. ";
-    echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
+    if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
+        echo "Success";
+    } else {
+        $error = mysqli_error($conn);
+        echo "There was a problem while deleting the data. ";
+        echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
+    }
 }
-
 
 mysqli_close($conn);
