+++ delete/medication.php
@@ -32,26 +32,29 @@ $enc_id="0x".bin2hex(openssl_encrypt($id, $cipher, $encryption_key, 0, $iv));
 
 mysqli_close($connect);
 
-$sql = "DELETE FROM `Medication`
-    WHERE
-        `id` = $enc_id
-        AND `medication` = '$medication'
-        AND `start` = '$start'
-        AND `stop` = '$stop'
-        AND `reason` = '$reason'
-        AND `intent` = '$intent'
-        AND `comment` = '$comment'";
-echo $sql;
-$sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
-VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
+$hasAdminRole = in_array("admin", explode(",", strtolower($roles)));
 
-if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
-    echo "Success";
-} else {
-    $error = mysqli_error($conn);
-    echo "There was a problem while saving the data. ";
-    echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
+if ($hasAdminRole) {
+    $sql = "DELETE FROM `Medication`
+        WHERE
+            `id` = $enc_id
+            AND `medication` = '$medication'
+            AND `start` = '$start'
+            AND `stop` = '$stop'
+            AND `reason` = '$reason'
+            AND `intent` = '$intent'
+            AND `comment` = '$comment'";
+    echo $sql;
+    $sql2 = "INSERT INTO `tracking`(`trackingid`, `username`, `email`, `roles`, `ip`, `date`)
+    VALUES ('$tracking','$username','$email','$roles','$ip','$datesystem')";
+
+    if (mysqli_query($conn, $sql) && mysqli_query($conn, $sql2)) {
+        echo "Success";
+    } else {
+        $error = mysqli_error($conn);
+        echo "There was a problem while saving the data. ";
+        echo "Please contact the admin of the site - Nadia Znassi. Your reference: ". $tracking .":". $error;
+    }
 }
-
 
 mysqli_close($conn);
